requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.137
marshmallow==3.23.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyMuPDF==1.24.12
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.16.0
